backend/cars/urls.py
from django.urls import include, path
from rest_framework import routers
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView
)
from .views import CarViewSet, DriverViewSet

app_name = 'cars'

router = routers.DefaultRouter()
router.register(r'cars', CarViewSet)
router.register(r'profile', DriverViewSet)

urlpatterns = [
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('', include(router.urls)),
]
