backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2024.2
sqlparse==0.4.4
typing_extensions==4.7.1
